Create QuakeMapFlipperV2.py
After seeing it declined to make texture adjustments in the first go, I prompted with this:

Please also include the texture adjustments you declined to include in point 5 of your reasoning. I understand it may not be perfect, but would prefer for more of it to look correct by performing the adjustments.