Create QuakeMapFlipperV3.py
Asked the following for clarification. It gave me this code as a response:

A reminder, the doors and other moving objects are built into the maps. The movement of those objects will also need to be adjusted.

Also, entities have a direction set in the map file. That direction indicates which direction is "forward". So, an enemy would face the direction. It is important this is properly changes as well. If it isn't, they may not "see" the player and react as they should.

Was that already accounted for?